mpesa-bookkeeping-copilot/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-cors-headers>=4.0
python-dotenv>=1.0
celery>=5.3
redis>=4.5
PyPDF2>=3.0.0
